+++ exportplan/tests/test_views.py
@@ -52,6 +52,7 @@ def test_export_plan_retrieve(authed_client, export_plan):
     response = authed_client.get(url)
 
     data = {
+        'name': None,
         'company': export_plan.company.id,
         'sso_id': export_plan.sso_id,
         'export_commodity_codes': export_plan.export_commodity_codes,
@@ -125,6 +126,7 @@ def test_export_plan_retrieve(authed_client, export_plan):
             }
         ],
         'pk': export_plan.pk,
+        'created': export_plan.created.isoformat().replace('+00:00', 'Z'),
     }
     assert response.status_code == 200
     assert response.json() == data
@@ -177,6 +179,7 @@ def test_export_plan_list_detail(authed_client, authed_supplier):
     response = authed_client.get(reverse('export-plan-detail-list'))
     assert list(response.json()[0].keys()) == [
         'pk',
+        'name',
         'created',
         'ui_progress',
         'export_countries',